Drop 'active' key from route descriptions
The 'active' key of route descriptions was previously used to determine
the checkbox status of routes in the UI. We now infer the status of
routes from their 'actions' key as returned by the Mailgun API so as to
minimize the risk of divergent metadata. To that end, we also adopt the
data structure layout for routes as Mailgun defines them, rather than
wrapping the data ourselves.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>