Drop background page
In the past, we used to fetch the initial routes from mailgun in the
(persistent) background page, and then poll the API every minute to
update the routes in case they are modified via the website, the API,
etc. This meant that we would continually poll the mailgun API. We now
query the API once on every load of the popup. While this introduces a
slight delay in the UI initialization (something we'll smooth over with
a loading indicator in the future), it minimizes the number of calls to
the mailgun API in the long run, and guarantees that routes are always
up-to-date.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>