Fix race-condition in npm 'dev' command
Previously, we would run 'webpack' and 'web-ext' in parallel in
conjunction with the 'clean-webpack-plugin'. This would sometimes cause
a race condition where the 'manifest.json' file wasn't copied to the
'addon' directory in time for 'web-ext' to pick it up. To avoid this, we
now never clean the 'addon' directory automatically, and make sure we
run 'webpack' once before launching 'webpack' and 'web-ext' in parallel
to automatically recompile and reload the extension.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>